Corrected name to permutator

+++ WinPredictor/Algos/SMOAlgo.cs
@@ -30,7 +30,7 @@ namespace WinPredictor.Algos
         public void Learn(IEnumerable<int> input, int output)
         {
             var inputForLearning = new double[][] { GetDoubleArray(input) };
-            _supportVectorMachine = _teacher.Learn(inputForLearning, new double[] { output });
+            _supportVectorMachine = _teacher.Learn(inputForLearning, new int[] { output });
         }
 
         private double[] GetDoubleArray(IEnumerable<int> input)
